requirements.txt
asn1crypto==0.24.0
cachetools==3.1.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
dialogflow==0.6.0
Flask==1.0.3
future==0.17.1
google-api-core==1.11.1
google-auth==1.6.3
googleapis-common-protos==1.6.0
grpcio==1.21.1
idna==2.8
infermedica-api==0.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pymongo==3.8.0
python-telegram-bot==11.1.0
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
telegram==0.0.1
urllib3==1.25.3
Werkzeug==0.15.4
